// src/types.ts
export type Role = 'ADMIN' | 'PARTICIPANT';

export interface User {
  id: string;
  identifier: string;
  role: Role;
}

export interface TeamMember {
  name: string;
  register_number: string;
  email?: string;
  phone?: string;
}

export interface Participant {
  id: string; // CM26-XXXX
  user_id: string;
  full_name: string;
  email: string;
  phone: string;
  register_number: string;
  department: string;
  year: string;
  team_name: string;
  team_members: TeamMember[];
  is_disabled: boolean;
  created_at: string;
}